refactor PersonalMessageManager using base MessageManager
